+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     - Fixed `button` view styles
     - Fixed `tabs` init value setting when `name` option is specified
     - Added `onInit` and `onChange` options for `tabs` view
+    - Extended `expanded` option for `tree` view, which can be a function now
+    - Added `onToggle` option for `tree` view
 
 ## 1.0.0-beta.10 (07-02-2019)